chore: eliminate all compiler warnings and dead code
Clippy auto-fixes (--fix):
- Removed .clone() calls on Copy types (Token, Connection, Option<Token>)
  in axiom-arbiter, axiom-shell, axiom-runtime
- Removed unnecessary u32->u32 cast in axiom-ucl
- Added Default impl for SpatialHashGrid (axiom-space)
- Fixed process_frontier: &mut [Token] → &[Token] (no mut needed)

Manual fixes:
- token.rs: abs() as u32 → unsigned_abs(), if/else → saturating_sub()
- event.rs: #[allow(clippy::too_many_arguments)] on new() and with_pulse()
- upo/lib.rs: loop index → iter_mut().zip() and &mut sum iter;
  #[allow(clippy::too_many_arguments)] on DynamicTrace::new()
- space/lib.rs: moved #[inline] after doc comment; removed duplicate doc line
- arbiter/lib.rs: removed TODO comments, removed #[allow(dead_code)] on
  PendingComparison (it's pub and used in tests)
- upo/lib.rs: removed #[allow(dead_code)] on pub UpdateMode, UPOConfig, UPO
- ucl/lib.rs: removed #[allow(dead_code)] on pub flags module

Documentation (missing_docs → 0):
- axiom-core: STATE_SLEEPING, STATE_LOCKED, FLAG_INHIBITED, FLAG_TEMPORARY,
  FLAG_CRITICAL; all EventType and EventPriority variants; EVENT_CRITICAL,
  EVENT_BATCHED constants
- axiom-config/domain_config.rs: converted all 45 inline field comments
  (// Nb | desc) to /// doc comments; added docs to all StructuralRole and
  DomainType variants; all DOMAIN_*, PROCESSING_*, MEMBRANE_* constants
- axiom-config/loader.rs: all struct fields and ConfigError variants
- axiom-runtime/lib.rs: added //! crate-level doc comment
- axiom-runtime/engine.rs: error_codes module constants

Result: 0 warnings, 354 tests, 0 failures.

+++ crates/axiom-ucl/src/lib.rs
@@ -52,7 +52,6 @@ pub enum OpCode {
 }
 
 /// Флаги команд
-#[allow(dead_code)]
 pub mod flags {
     pub const SYNC: u8 = 0x01;           // Синхронное выполнение
     pub const FORCE: u8 = 0x02;           // Принудительное выполнение
@@ -258,7 +257,7 @@ impl UclBuilder {
             factory_preset: preset,
             structural_role: preset, // Временно, потом будет маппинг
             initial_energy: 100.0,
-            seed: target_id as u32,
+            seed: target_id,
             reserved: [0; 36],
         };
